+++ protocol-dns/src/main/java/org/apache/directory/server/dns/io/encoder/ResourceRecordEncoder.java
@@ -120,7 +120,7 @@ public abstract class ResourceRecordEncoder implements RecordEncoder
 
     protected void putRecordClass( ByteBuffer byteBuffer, RecordClass recordClass )
     {
-        byteBuffer.putShort( ( short ) recordClass.convert() );
+        byteBuffer.putShort( recordClass.convert() );
     }
